Refactor: 서비스 메서드 위치를 더 적절한 곳으로 이동
